requirements.txt
# ipywidgets
# umap-learn==0.5.1
fastapi==0.95.2
uvicorn[standard]==0.22.0
transformers[torch,sentencepiece,vision,optuna,sklearn,onnxruntime]==4.41.1
sentencepiece==0.1.99
snscrape==0.5.0.20230113
matplotlib==3.9.4
pandas==2.3.3
python-multipart==0.0.6
pydantic==1.10.12
joblib==1.5.2
typing-extensions==4.15.0
pyarrow==15.0.2
praw==7.8.1
boto3==1.28.39
redis==4.5.5
kafka-python-ng==2.2.3
psutil==5.9.5
psycopg2-binary==2.9.7
optimum[onnxruntime]==1.16.2
onnxruntime==1.17.0
